jump to comment list when creating and deleting comment
