+++ app/views/tracks/show.html.erb
@@ -75,7 +75,7 @@
       <%= render partial: "tracks/track", locals: { track: t } %>
     <% end %>
   </section>
-  <section class="w-full">
+  <section id="comments" class="w-full">
     <h1 class="font-bold text-lg mb-2"><%= t("label.comments") %></h1>
     <div class="flex justify-start items-stretch gap-4 my-8">
       <%= render partial: "users/shared/avatar", locals: { size: 10 } %>
@@ -85,9 +85,9 @@
         url: comments_path,
         method: :post,
         data: {
-          controller: "comment-form",
-          comment_form_target: "form",
-          action: "turbo:submit-end->comment-form#submitForm",
+          controller: "comment-create",
+          comment_create_target: "form",
+          action: "turbo:submit-end->comment-create#submitForm",
         }) do |f| %>
         <%= f.hidden_field :entity_type, value: Track.name %>
         <%= f.hidden_field :entity_id, value: @track.id %>
@@ -122,7 +122,7 @@
     </div>
 
     <div>
-      <%= turbo_frame_tag "track-comments", target: "_top", data: { turbo_cache: false } do %>
+      <%= turbo_frame_tag "track-comments", target: "_top", data: { turbo_temporary: false } do %>
         <%= render partial: "comments/list", locals: { comments: @comments } %>
       <% end %>
